# docs/openapi.json
{
 "openapi": "3.0.1",
 "info": {
  "title": "OpenAPI definition",
  "version": "v0"
 },
 "servers": [
  {
   "url": "http://localhost:18090",
   "description": "Generated server url"
  }
 ],
 "paths": {
...
  "/admin/vote/options": {
   "get": {
    "tags": [
     "engine-admin-controller"
    ],
    "operationId": "voteOptions",
    "parameters": [
     {
      "name": "serviceId",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "engine-admin-controller"
    ],
    "operationId": "addOption",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OptionReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/services": {
   "get": {
    "tags": [
     "admin-controller"
    ],
    "operationId": "services_2",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     },
     {
      "name": "size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "admin-controller"
    ],
    "operationId": "createService",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ServiceReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/services/{id}/close": {
   "post": {
    "tags": [
     "engine-admin-controller"
    ],
    "operationId": "close",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/quiz/questions": {
   "get": {
    "tags": [
     "engine-admin-controller"
    ],
    "operationId": "quizQuestions",
    "parameters": [
     {
      "name": "serviceId",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "engine-admin-controller"
    ],
    "operationId": "addQuestion",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/QuestionReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
  "/admin/partners": {
   "get": {
    "tags": [
     "admin-controller"
    ],
    "operationId": "partners",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     },
     {
      "name": "size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "admin-controller"
    ],
    "operationId": "createPartner",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PartnerReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/content/items": {
   "get": {
    "tags": [
     "engine-admin-controller"
    ],
    "operationId": "contentItems",
    "parameters": [
     {
      "name": "serviceId",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "engine-admin-controller"
    ],
    "operationId": "addItem",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ItemReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
  "/admin/billing/periods": {
   "get": {
    "tags": [
     "billing-query-controller"
    ],
    "operationId": "periods",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "billing-controller"
    ],
    "operationId": "close_1",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PeriodReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
  "/admin/billing/payouts": {
   "get": {
    "tags": [
     "billing-query-controller"
    ],
    "operationId": "payouts_1",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "billing-controller"
    ],
    "operationId": "createPayout",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PayoutReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
  "/admin/billing/payouts/{id}/pay": {
   "post": {
    "tags": [
     "billing-controller"
    ],
    "operationId": "pay",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PayReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
  "/admin/billing/payouts/{id}/cancel": {
   "post": {
    "tags": [
     "billing-controller"
    ],
    "operationId": "cancel",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
  "/admin/billing/adjustments": {
   "post": {
    "tags": [
     "billing-controller"
    ],
    "operationId": "adjust",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AdjustReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/vote/options/{id}": {
   "delete": {
    "tags": [
     "engine-admin-controller"
    ],
    "operationId": "deleteOption",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK"
     }
    }
   }
  },
...
  "/admin/quiz/questions/{id}": {
   "delete": {
    "tags": [
     "engine-admin-controller"
    ],
    "operationId": "deleteQuestion",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK"
     }
    }
   }
  },
...
  "/admin/content/items/{id}": {
   "delete": {
    "tags": [
     "engine-admin-controller"
    ],
    "operationId": "deleteItem",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK"
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "OptionReq": {
    "type": "object",
    "properties": {
     "serviceId": {
      "type": "integer",
      "format": "int64"
     },
     "code": {
      "type": "string"
     },
     "label": {
      "type": "string"
     }
    }
   },
...
   "ServiceReq": {
    "type": "object",
    "properties": {
     "name": {
      "type": "string"
     },
     "type": {
      "type": "string",
      "enum": [
       "VOTE",
       "QUIZ",
       "PREMIUM_CONTENT",
       "SUBSCRIPTION",
       "ALERT"
      ]
     },
     "partnerId": {
      "type": "integer",
      "format": "int64"
     },
     "shortCodeId": {
      "type": "integer",
      "format": "int64"
     },
     "regulated": {
      "type": "boolean"
     },
     "consentMode": {
      "type": "string",
      "enum": [
       "SIMPLE_OPT_IN",
       "DOUBLE_OPT_IN",
       "API_ACTIVATION"
      ]
     },
     "opensAt": {
      "type": "string",
      "format": "date-time"
     },
     "closesAt": {
      "type": "string",
      "format": "date-time"
     },
     "maxActionsPerMsisdn": {
      "type": "integer",
      "format": "int32"
     },
     "defaultLang": {
      "type": "string"
     },
     "replyOk": {
      "type": "string"
     },
     "replyStop": {
      "type": "string"
     },
     "replyHelp": {
      "type": "string"
     },
     "replyLimit": {
      "type": "string"
     },
     "replyClosed": {
      "type": "string"
     },
     "maxTps": {
      "type": "integer",
      "format": "int32"
     }
    }
   },
...
   "QuestionReq": {
    "type": "object",
    "properties": {
     "serviceId": {
      "type": "integer",
      "format": "int64"
     },
     "position": {
      "type": "integer",
      "format": "int32"
     },
     "question": {
      "type": "string"
     },
     "answers": {
      "type": "string"
     },
     "points": {
      "type": "integer",
      "format": "int32"
     },
     "replyCorrect": {
      "type": "string"
     },
     "replyWrong": {
      "type": "string"
     }
    }
   },
   "PartnerReq": {
    "type": "object",
    "properties": {
     "name": {
      "type": "string"
     },
     "sharePercent": {
      "type": "number"
     },
     "webhookUrl": {
      "type": "string"
     },
     "webhookSecret": {
      "type": "string"
     },
     "maxTps": {
      "type": "integer",
      "format": "int32"
     }
    }
   },
...
   "ItemReq": {
    "type": "object",
    "properties": {
     "serviceId": {
      "type": "integer",
      "format": "int64"
     },
     "code": {
      "type": "string"
     },
     "title": {
      "type": "string"
     },
     "body": {
      "type": "string"
     },
     "url": {
      "type": "string"
     },
     "maxUses": {
      "type": "integer",
      "format": "int32"
     },
     "ttlHours": {
      "type": "integer",
      "format": "int32"
     }
    }
   },
   "PeriodReq": {
    "type": "object",
    "properties": {
     "from": {
      "type": "string",
      "format": "date-time"
     },
     "to": {
      "type": "string",
      "format": "date-time"
     }
    }
   },
   "PayoutReq": {
    "type": "object",
    "properties": {
     "partnerId": {
      "type": "integer",
      "format": "int64"
     },
     "from": {
      "type": "string",
      "format": "date-time"
     },
     "to": {
      "type": "string",
      "format": "date-time"
     }
    }
   },
   "PayReq": {
    "type": "object",
    "properties": {
     "reference": {
      "type": "string"
     }
    }
   },
   "AdjustReq": {
    "type": "object",
    "properties": {
     "eventId": {
      "type": "string"
     },
     "reason": {
      "type": "string"
     }
    }
   },
...
  }
 }
}